+++ tests/test_smiles.py
@@ -36,6 +36,27 @@ def test_to_canonical_smiles(AceticAcid):
     assert smiles == correct
 
 
+def test_to_canonical_smiles_openbabel(AceticAcid):
+    """Create a SMILES string from a system"""
+    correct = "CC(=O)O"
+    smiles = AceticAcid.to_smiles(canonical=True, flavor="openbabel")
+
+    if smiles != correct:
+        print(smiles)
+    assert smiles == correct
+
+
+@pytest.mark.openeye
+def test_to_canonical_smiles_openeye(AceticAcid):
+    """Create a SMILES string from a system"""
+    correct = "CC(=O)O"
+    smiles = AceticAcid.to_smiles(canonical=True, flavor="openeye")
+
+    if smiles != correct:
+        print(smiles)
+    assert smiles == correct
+
+
 def test_several_molecules(CH3COOH_3H2O):
     """System with acetic acid and 3 waters"""
     system = CH3COOH_3H2O
@@ -70,3 +91,31 @@ def test_from_smiles(configuration):
         print(result)
 
     assert result == correct
+    assert configuration.name == "acetic acid"
+
+
+def test_from_smiles_openbabel(configuration):
+    """Create a configuration from SMILES"""
+    correct = "CC(=O)O"
+    configuration.from_smiles("OC(=O)C", name="acetic acid", flavor="openbabel")
+    result = configuration.to_smiles(canonical=True)
+
+    if result != correct:
+        print(result)
+
+    assert result == correct
+    assert configuration.name == "acetic acid"
+
+
+@pytest.mark.openeye
+def test_from_smiles_openeye(configuration):
+    """Create a configuration from SMILES"""
+    correct = "CC(=O)O"
+    configuration.from_smiles("OC(=O)C", name="acetic acid", flavor="openeye")
+    result = configuration.to_smiles(canonical=True)
+
+    if result != correct:
+        print(result)
+
+    assert result == correct
+    assert configuration.name == "acetic acid"